Segundo commit

+++ app/Http/Controllers/Posts/PostController.php
@@ -32,9 +32,7 @@ class PostController extends Controller
 
     public function destroy(Post $post){
         
-        if(!$post->ownedBy(auth()->user())){
-            
-        }
+        $this->authorize('delete', $post);
 
         $post->delete();
